+++ x/gamm/keeper/multihop_test.go
@@ -2,6 +2,7 @@ package keeper_test
 
 import (
 	"fmt"
+	"github.com/osmosis-labs/osmosis/v11/x/gamm/pool-models/balancer"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
@@ -16,9 +17,10 @@ func (suite *KeeperTestSuite) TestBalancerPoolSimpleMultihopSwapExactAmountIn() 
 	}
 
 	tests := []struct {
-		name       string
-		param      param
-		expectPass bool
+		name              string
+		param             param
+		expectPass        bool
+		reducedFeeApplied bool
 	}{
 		{
 			name: "Proper swap - foo -> bar(pool 1) - bar(pool 2) -> baz",
@@ -38,6 +40,25 @@ func (suite *KeeperTestSuite) TestBalancerPoolSimpleMultihopSwapExactAmountIn() 
 			},
 			expectPass: true,
 		},
+		{
+			name: "Swap - foo -> uosmo(pool 1) - uosmo(pool 2) -> baz with a half fee applied",
+			param: param{
+				routes: []types.SwapAmountInRoute{
+					{
+						PoolId:        1,
+						TokenOutDenom: "uosmo",
+					},
+					{
+						PoolId:        2,
+						TokenOutDenom: "baz",
+					},
+				},
+				tokenIn:           sdk.NewCoin("foo", sdk.NewInt(100000)),
+				tokenOutMinAmount: sdk.NewInt(1),
+			},
+			reducedFeeApplied: true,
+			expectPass:        true,
+		},
 	}
 
 	for _, test := range tests {
@@ -46,8 +67,14 @@ func (suite *KeeperTestSuite) TestBalancerPoolSimpleMultihopSwapExactAmountIn() 
 
 		suite.Run(test.name, func() {
 			// Prepare 2 pools
-			suite.PrepareBalancerPool()
-			suite.PrepareBalancerPool()
+			suite.PrepareBalancerPoolWithPoolParams(balancer.PoolParams{
+				SwapFee: sdk.NewDecWithPrec(1, 2),
+				ExitFee: sdk.NewDec(0),
+			})
+			suite.PrepareBalancerPoolWithPoolParams(balancer.PoolParams{
+				SwapFee: sdk.NewDecWithPrec(1, 2),
+				ExitFee: sdk.NewDec(0),
+			})
 
 			keeper := suite.App.GAMMKeeper
 
@@ -69,6 +96,20 @@ func (suite *KeeperTestSuite) TestBalancerPoolSimpleMultihopSwapExactAmountIn() 
 					}
 					return dec
 				}()
+
+				calcOutAmountAsSeparateSwaps := func(feeMultiplier sdk.Dec) sdk.Coin {
+					cacheCtx, _ := suite.Ctx.CacheContext()
+					nextTokenIn := test.param.tokenIn
+					for _, hop := range test.param.routes {
+						tokenOut, err := keeper.SwapExactAmountIn(cacheCtx, suite.TestAccs[0], hop.PoolId, nextTokenIn, hop.TokenOutDenom, sdk.OneInt(), feeMultiplier)
+						suite.Require().NoError(err)
+						nextTokenIn = sdk.NewCoin(hop.TokenOutDenom, tokenOut)
+					}
+					return nextTokenIn
+				}
+
+				tokenOutCalculatedAsSeparateSwaps := calcOutAmountAsSeparateSwaps(sdk.OneDec())
+				tokenOutCalculatedAsSeparateSwapsWithoutFee := calcOutAmountAsSeparateSwaps(sdk.ZeroDec())
 
 				tokenOutAmount, err := keeper.MultihopSwapExactAmountIn(suite.Ctx, suite.TestAccs[0], test.param.routes, test.param.tokenIn, test.param.tokenOutMinAmount)
 				suite.NoError(err, "test: %v", test.name)
@@ -95,6 +136,14 @@ func (suite *KeeperTestSuite) TestBalancerPoolSimpleMultihopSwapExactAmountIn() 
 				// Ratio of the token out should be between the before spot price and after spot price.
 				sp := test.param.tokenIn.Amount.ToDec().Quo(tokenOutAmount.ToDec())
 				suite.True(sp.GT(spotPriceBefore) && sp.LT(spotPriceAfter), "test: %v", test.name)
+
+				if test.reducedFeeApplied {
+					suite.Require().True(tokenOutAmount.GT(tokenOutCalculatedAsSeparateSwaps.Amount))
+					suite.Require().True(tokenOutAmount.LTE(tokenOutCalculatedAsSeparateSwapsWithoutFee.Amount))
+				} else {
+					suite.Require().True(tokenOutAmount.Equal(tokenOutCalculatedAsSeparateSwaps.Amount))
+				}
+
 			} else {
 				_, err := keeper.MultihopSwapExactAmountIn(suite.Ctx, suite.TestAccs[0], test.param.routes, test.param.tokenIn, test.param.tokenOutMinAmount)
 				suite.Error(err, "test: %v", test.name)
@@ -111,9 +160,10 @@ func (suite *KeeperTestSuite) TestBalancerPoolSimpleMultihopSwapExactAmountOut()
 	}
 
 	tests := []struct {
-		name       string
-		param      param
-		expectPass bool
+		name              string
+		param             param
+		expectPass        bool
+		reducedFeeApplied bool
 	}{
 		{
 			name: "Proper swap: foo -> bar (pool 1), bar -> baz (pool 2)",
@@ -133,6 +183,25 @@ func (suite *KeeperTestSuite) TestBalancerPoolSimpleMultihopSwapExactAmountOut()
 			},
 			expectPass: true,
 		},
+		{
+			name: "Swap - foo -> uosmo(pool 1) - uosmo(pool 2) -> baz with a half fee applied",
+			param: param{
+				routes: []types.SwapAmountOutRoute{
+					{
+						PoolId:       1,
+						TokenInDenom: "foo",
+					},
+					{
+						PoolId:       2,
+						TokenInDenom: "uosmo",
+					},
+				},
+				tokenInMaxAmount: sdk.NewInt(90000000),
+				tokenOut:         sdk.NewCoin("baz", sdk.NewInt(100000)),
+			},
+			expectPass:        true,
+			reducedFeeApplied: true,
+		},
 	}
 
 	for _, test := range tests {
@@ -141,8 +210,14 @@ func (suite *KeeperTestSuite) TestBalancerPoolSimpleMultihopSwapExactAmountOut()
 
 		suite.Run(test.name, func() {
 			// Prepare 2 pools
-			suite.PrepareBalancerPool()
-			suite.PrepareBalancerPool()
+			suite.PrepareBalancerPoolWithPoolParams(balancer.PoolParams{
+				SwapFee: sdk.NewDecWithPrec(1, 2),
+				ExitFee: sdk.NewDec(0),
+			})
+			suite.PrepareBalancerPoolWithPoolParams(balancer.PoolParams{
+				SwapFee: sdk.NewDecWithPrec(1, 2),
+				ExitFee: sdk.NewDec(0),
+			})
 
 			keeper := suite.App.GAMMKeeper
 
@@ -165,6 +240,21 @@ func (suite *KeeperTestSuite) TestBalancerPoolSimpleMultihopSwapExactAmountOut()
 					}
 					return dec
 				}()
+
+				calcInAmountAsSeparateSwaps := func(feeMultiplier sdk.Dec) sdk.Coin {
+					cacheCtx, _ := suite.Ctx.CacheContext()
+					nextTokenOut := test.param.tokenOut
+					for i := len(test.param.routes) - 1; i >= 0; i-- {
+						hop := test.param.routes[i]
+						tokenOut, err := keeper.SwapExactAmountOut(cacheCtx, suite.TestAccs[0], hop.PoolId, hop.TokenInDenom, sdk.NewInt(100000000), nextTokenOut, feeMultiplier)
+						suite.Require().NoError(err)
+						nextTokenOut = sdk.NewCoin(hop.TokenInDenom, tokenOut)
+					}
+					return nextTokenOut
+				}
+
+				tokenInCalculatedAsSeparateSwaps := calcInAmountAsSeparateSwaps(sdk.OneDec())
+				tokenInCalculatedAsSeparateSwapsWithoutFee := calcInAmountAsSeparateSwaps(sdk.ZeroDec())
 
 				tokenInAmount, err := keeper.MultihopSwapExactAmountOut(suite.Ctx, suite.TestAccs[0], test.param.routes, test.param.tokenInMaxAmount, test.param.tokenOut)
 				suite.Require().NoError(err, "test: %v", test.name)
@@ -193,6 +283,13 @@ func (suite *KeeperTestSuite) TestBalancerPoolSimpleMultihopSwapExactAmountOut()
 				sp := tokenInAmount.ToDec().Quo(test.param.tokenOut.Amount.ToDec())
 				fmt.Printf("spBefore %s, spAfter %s, sp actual %s\n", spotPriceBefore, spotPriceAfter, sp)
 				suite.True(sp.GT(spotPriceBefore) && sp.LT(spotPriceAfter), "multi-hop spot price wrong, test: %v", test.name)
+
+				if test.reducedFeeApplied {
+					suite.Require().True(tokenInAmount.LT(tokenInCalculatedAsSeparateSwaps.Amount))
+					suite.Require().True(tokenInAmount.GTE(tokenInCalculatedAsSeparateSwapsWithoutFee.Amount))
+				} else {
+					suite.Require().True(tokenInAmount.Equal(tokenInCalculatedAsSeparateSwaps.Amount))
+				}
 			} else {
 				_, err := keeper.MultihopSwapExactAmountOut(suite.Ctx, suite.TestAccs[0], test.param.routes, test.param.tokenInMaxAmount, test.param.tokenOut)
 				suite.Error(err, "test: %v", test.name)