+++ x/txfees/types/expected_keepers.go
@@ -20,6 +20,7 @@ type GammKeeper interface {
 		tokenIn sdk.Coin,
 		tokenOutDenom string,
 		tokenOutMinAmount sdk.Int,
+		swapFeeMultiplier sdk.Dec,
 	) (tokenOutAmount sdk.Int, err error)
 }